+++ fiboa_cli/const.py
@@ -7,5 +7,6 @@ LOG_STATUS_COLOR = {
 
 SUPPORTED_PROTOCOLS = ["http", "https", "s3", "gs"]
 
-STAC_COLLECTION_SCHEMA = "http://schemas.stacspec.org/v1.0.0/collection-spec/json-schema/collection.json"
+STAC_COLLECTION_SCHEMA = "http://schemas.stacspec.org/v{version}/collection-spec/json-schema/collection.json"
+GEOPARQUET_SCHEMA = "https://geoparquet.org/releases/v{version}/schema.json"
 STAC_TABLE_EXTENSION = "https://stac-extensions.github.io/table/v1.2.0/schema.json"